Modified sim to send accurate data to the server
The server now receives pairwise distances for beacons. These
measurements are "identical" to those that would be received by the
server in practice, assuming nothing is moving. Right now the data is
just being stored in a global array, so I need to come up with a better
way of averaging the data and then figuring out global positions based
on all of the data.