Made some speed improvements to math test.
Also performed some interesting experiments. Turns out that the
initialization of the points is important, and random initializaiton
leads to different results every time. Instead, what I'll probably do is
use the previous best guess as a seed for the next round of gradient
descent. This should have the effect of keeping the world coordinate
frame relatively stable.

I also found out that experimenting with the learning rate can make or
break the program -- too slow of a learning rate and the program doesn't
converge in time, too fast and it starts throwing NaNs for some reason.